Put every regolith surface on the same reflectance law
The terrain got the regolith BRDF last commit; the roads, hardstand, rubble and
ground stains did not, and that inconsistency is worth about 4x. A Lambertian
surface and a Hapke surface of identical albedo differ by roughly that factor at
the phase angles a polar scene lives at, so surfaces authored to look like the
ground now read as poured concrete beside it — pale, flat, and pasted on. The
first render made it obvious.

GRADED SURFACES (BaseRoads: road crust, hardstand, spoil, boulders). Swapped
three's diffuse lobe for the regolith BRDF and added the same shadow fill the
terrain uses, so a road in shadow can no longer sit at a different depth from the
ground it crosses. Everything else in three's light loop is kept, most importantly
the shadow attenuation that arrives folded into directLight.color, and the
surfaces keep their own albedo and normal maps — only the law changed. Dropping
the GGX specular lobe is intended: regolith has no highlight worth the name, these
run at roughness 0.96-1.0, and the terrain beside them has no specular term.

STAINS (GroundDisturbance). These were absolute colours on an unlit material,
which was correct against a baked ground and inverted the moment the ground became
computed: they held fixed brightness into shadow, where a mark on soil cannot be
brighter than the soil, and they flattened the per-pixel relief they covered. A
stain physically IS a patch of ground with a different albedo, so that is what it
became — a multiplicative factor, via MultiplyBlending, over ground that has
already been lit. The authored tones are reinterpreted as ratios against the
brightness they were originally picked against, which is the part of the artwork
still worth keeping.

That multiply is exact here rather than approximate, and the reason is a pipeline
property worth recording: the tone curve is an EffectComposer pass, so the scene
renders into a linear HalfFloatType buffer with no clamp. In an 8-bit sRGB target
the same blend would understate a stain by 0.88 stops over ground at 0.005, and
the blast halo could not brighten at all. Both numbers are in the comment.

Shared derivations rather than parallel ones, since this bug has now appeared
three times from the same substitution. litGroundRadiance and shadowFillRadiance
move into regolith.ts as functions of the sun, which also removes the scene
coupling from that module and is the shape a moving sun will need. The BRDF swap
becomes a factory parameterised over three's material struct, so the terrain's
Lambert version and the surfaces' Standard version cannot drift — a test asserts
their bodies are character-for-character identical.

Renames terrainShader.ts to regolithShader.ts, which is what it now is.

31 new tests, including the graded and stain patch sets against three's real
Standard and Basic sources. Writing them caught a fourth 4x error, in the shadow
fill, before it shipped.
