+++ ui/scripts/.mocharc-cypress-unit.json
@@ -32,7 +32,7 @@
     "cypress/integration/unit/lunar-atlas-selectors.cy.ts",
     "cypress/integration/unit/lunar-atlas-shading.cy.ts",
     "cypress/integration/unit/lunar-atlas-southpole.cy.ts",
-    "cypress/integration/unit/lunar-atlas-terrain-shader.cy.ts",
+    "cypress/integration/unit/lunar-atlas-regolith-shader.cy.ts",
     "cypress/integration/unit/team-roles.cy.tsx",
     "cypress/integration/unit/renew-subscription.cy.ts",
     "cypress/integration/unit/marketplace-usdc-onramp.cy.tsx",
